+++ frontend/packages/data-portal/app/components/Deposition/MethodLinks/type.ts
@@ -3,7 +3,7 @@ import { MethodType } from 'app/constants/methodTypes'
 // note: below is also used for ordering in lists
 export const METHOD_LINK_TYPES = [
   'source_code',
-  'model_weights',
+  'models_weights',
   'website',
   'documentation',
   'other',