+++ frontend/packages/data-portal/app/graphql/getDatasetsDiffer.ts
@@ -9,7 +9,7 @@ import {
   GetDatasetsV2Query,
 } from 'app/__generated_v2__/graphql'
 
-import { convertReconstructionMethodToV2 } from './common'
+import { convertReconstructionMethodToV2, removeTypenames } from './common'
 
 /* eslint-disable no-console, no-param-reassign */
 export function logIfHasDiff(
@@ -18,9 +18,12 @@ export function logIfHasDiff(
   v1FilterValues: GetDatasetsFilterDataQuery,
   v2: GetDatasetsV2Query,
 ): void {
-  console.log('Checking for datasets query diffs')
+  console.log(
+    `Checking for datasets query diffs ${new Date().toLocaleString()}`,
+  )
 
   v2 = structuredClone(v2)
+  removeTypenames(v2)
 
   // Counts not used.
   // Create consistent sort order.