+++ frontend/packages/data-portal/app/graphql/getDepositionByIdV2.server.ts
@@ -51,32 +51,6 @@ const GET_DEPOSITION_BY_ID = gql(`
           count
         }
       }
-      distinctOrganismNames: datasetsAggregate {
-        aggregate {
-          count
-          groupBy {
-            organismName
-          }
-        }
-      }
-      distinctObjectNames: annotationsAggregate {
-        aggregate {
-          count
-          groupBy {
-            objectName
-          }
-        }
-      }
-      distinctShapeTypes: annotationsAggregate {
-        aggregate {
-          count
-          groupBy {
-            annotationShapes {
-              shapeType
-            }
-          }
-        }
-      }
       annotationMethodCounts: annotationsAggregate {
         aggregate {
           count
@@ -179,7 +153,11 @@ export async function getDepositionByIdV2({
         filterState: getFilterState(params),
         depositionId: id,
       }),
-      datasetsByDepositionFilter: depositionIdFilter,
+      datasetsByDepositionFilter: {
+        runs: {
+          annotations: depositionIdFilter,
+        },
+      },
       tiltseriesByDepositionFilter: depositionIdFilter,
       tomogramsByDepositionFilter: depositionIdFilter,
       annotationsByDepositionFilter: depositionIdFilter,